feat: 增强用户认证与查询功能，添加日志记录，优化异常处理

+++ admin-backend/admin-shared/admin-shared-common/src/main/java/io/github/faustofan/admin/shared/common/dto/PageResponse.java
@@ -13,10 +13,10 @@ import java.util.List;
  * @param <T>           type of items in the list
  */
 public record PageResponse<T>(
-            Integer totalElements,
-            Integer totalPages,
-            Integer currentPage,
-            Integer pageSize,
+            long totalElements,
+            long totalPages,
+            long currentPage,
+            long pageSize,
             List<T> items
 ) {
 }
